Moved old kai_config.py into kai_solution_server/
